% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{contextualDistance}
\alias{contextualDistance}
\title{contextualDistance}
\usage{
contextualDistance(
  word_a,
  word_b,
  gallitoCode,
  spaceName,
  neutral_state = TRUE,
  state,
  min_cosine_contour = 0.3,
  neighbors = 100
)
}
\arguments{
\item{word_a}{The first word the function will evaluate.}

\item{word_b}{The second word the function will evaluate.}

\item{gallitoCode}{Gallito API password to extract information from the LSA semantic space.}

\item{spaceName}{Gallito API LSA semantic space you want to use}

\item{neutral_state}{By default the function will take the neutral state between the two words subspaces
to evaluate the similarity}

\item{state}{The initial state the function will use to calculate the similarity between the two words in case
`neutral_state = FALSE`.}

\item{min_cosine_contour}{minimum cosine required for being a contour exemplar. By default `min_cosine_contour = 0.3`.}

\item{neighbors}{The number of neighbors inside the contour of the word. By default `neighbors = 100`.}
}
\value{
The function will return a value between 0 and +Inf indicating the distance between the two words.
}
\description{
contextualDistance is a function that takes two words and a state vector and returns the contextual
distance between them. To see more details about the contextual distance go to Gabora and Aerts (2002).
}
