% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{gallitoContour}
\alias{gallitoContour}
\title{gallitoContour}
\usage{
gallitoContour(
  word,
  gallitoCode,
  spaceName,
  neighbors = 100,
  min_cosine_contour = 0.3
)
}
\arguments{
\item{word}{A string or a tupple of strings separated by "_" indicating the word for which you want to
extract the contour}

\item{gallitoCode}{Gallito API password to extract information from the LSA semantic space.}

\item{spaceName}{Gallito API LSA semantic space you want to use.}

\item{neighbors}{The number of neighbors inside the contour of the word. By default `neighbors = 100`.}

\item{min_cosine_contour}{minimum cosine required for being a contour exemplar. By default `min_cosine_contour = 0.3`.}
}
\value{
A data frame with the contour of the word is returned.
}
\description{
gallitoContour is a function that will extract a word contour from Gallito API LSA semantic space. It takes
a word or a tupple (separated by "_") and returns a data frame with the contour of that word.
}
