% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{multidimensionalProjector}
\alias{multidimensionalProjector}
\title{multidimensionalProjector}
\usage{
multidimensionalProjector(subspace)
}
\arguments{
\item{subspace}{A multidimensional subspace in a matrix format.}
}
\value{
The function will return a matrix with the projector of the subspace.
}
\description{
multidimensionalProjector is a function that takes a multidimensional subspace and calculates its projector using
the sum of the outter products of its vectors. You can use the `subspaceGeneration` function from `QLSA` library to
define a subspace in order to calculate its multidimensional projector
}
