% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{neutralState}
\alias{neutralState}
\title{neutralState}
\usage{
neutralState(word_a, word_b, gallitoCode, spaceName)
}
\arguments{
\item{word_a}{The first word the function will evaluate.}

\item{word_b}{The second word the function will evaluate.}

\item{gallitoCode}{Gallito API password to extract information from the LSA semantic space.}

\item{spaceName}{Gallito API LSA semantic space you want to use}
}
\value{
The function will return a value between 0 and 1 indicating the similarity between the two words.
}
\description{
neutralState is a function that takes two words projectors and estimates the neutral state
vector in between. To see more details about the quantum similarity go to Pothos and Busemeyer (2011).
}
