% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{neutralStateProjectors}
\alias{neutralStateProjectors}
\title{neutralStateProjectors}
\usage{
neutralStateProjectors(word_a_vec, word_b_vec, Projector_a, Projector_b)
}
\arguments{
\item{word_a_vec}{The first word one-dimensional vector.}

\item{word_b_vec}{The second word one-dimensional vector.}

\item{Projector_a}{The first word projector.}

\item{Projector_b}{The second word projector.}
}
\value{
The function will return a value between 0 and 1 indicating the similarity between the two words.
}
\description{
neutralStateProjectors is a function that takes two words projectors and estimates the neutral state
vector in between. To see more details about the quantum similarity go to Pothos and Busemeyer (2011).
}
