% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{quantumSimilarity}
\alias{quantumSimilarity}
\title{quantumSimilarity}
\usage{
quantumSimilarity(word_a, word_b, gallitoCode, neutral_state = TRUE, state)
}
\arguments{
\item{word_a}{The first word the function will evaluate.}

\item{word_b}{The second word the function will evaluate.}

\item{gallitoCode}{Gallito API password to extract information from the LSA semantic space.}

\item{neutral_state}{By default the function will take the neutral state between the two words subspaces
to evaluate the similarity}

\item{state}{The initial state the function will use to calculate the similarity between the two words in case
`neutral_state = FALSE`.}
}
\value{
The function will return a value between 0 and 1 indicating the similarity between the two words.
}
\description{
quantumSimilarity is a function that takes two words projectors and a state vector and returns the quantum
similarity between them. To see more details about the quantum similarity go to Pothos and Busemeyer (2011).
}
