% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{subspaceGeneration}
\alias{subspaceGeneration}
\title{subspaceGeneration}
\usage{
subspaceGeneration(
  word,
  gallitoCode,
  spaceName,
  min_cosine = 0.5,
  min_reilability = 0.85
)
}
\arguments{
\item{word}{A string or a tupple of strings separated by "_" indicating the word for which you want to
define the subspace.}

\item{gallitoCode}{Gallito API password to extract information from the LSA semantic space.}

\item{spaceName}{Gallito API LSA semantic space you want to use.}

\item{min_cosine}{The minimum cosine the function will use to return the plots of similar words for
each new dimension of the subspace. By default `min_cosine = 0.5`.}

\item{min_reilability}{The minimum reilability the function will consider to decide
that a reorthogonalized dimension is equivalent to the dimension extracted in the factorial solution.
By default `min_reilability = 0.85`.}
}
\value{
The function will return a list with the subspace as `subspace`, the reilability test as `reilability_test`,
the subspace graphical information as `subspace_info` and the EFA results as `EFA_info`.
}
\description{
subspaceGeneration is a function that decides the number of dimensions a contour deserves using the `paran` package,
compute an EFA (Exploratory Factor Analysis) with that number of dimensions using the `psych` package,
and reorthogonalize the solution to define a basis for the new subspace using the `pracma` package.
}
