% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{unidimensionalProjector}
\alias{unidimensionalProjector}
\title{unidimensionalProjector}
\usage{
unidimensionalProjector(vector)
}
\arguments{
\item{vector}{A unidimensional subspace in a vector format.}
}
\value{
The function will return a matrix with the projector of the subspace.
}
\description{
unidimensionalProjector is a function that takes a unidimensional subspace and calculates its projector using
the outter product operation. You can use the `wordVector` function to extract a word vector and calculate its
unidimensional projector.
}
