% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLSA.R
\name{wordVector}
\alias{wordVector}
\title{wordVector}
\usage{
wordVector(word, gallitoCode)
}
\arguments{
\item{word}{A string or a tupple of strings separated by "_" indicating the word you want to extract.}

\item{gallitoCode}{Gallito API password to extract information from the LSA semantic space}
}
\value{
The word vector is returned
}
\description{
wordVector is a function that will extract a word vector from Gallito API LSA semantic space. It takes a word or a tupple (separated by "_") and returns the vector of that word in the LSA semanti space.
}
